# backend/internal/launchcode/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Facade Launch API
  version: 1.0.0
servers:
  - url: /
    description: Current LaunchServer origin
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Ant-Api-Key
      description: 应用设置中启用 Launch API Key 后必填
  schemas:
    ErrorResponse:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
        authHeader:
          type: string
          description: 鉴权失败时返回期望的 header 名
    Profile:
      type: object
      description: 实例完整快照
      properties:
        profileId:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        profileName:
          type: string
          example: buyer-001
        userDataDir:
          type: string
        coreId:
          type: string
        fingerprintArgs:
          type: array
          items:
            type: string
        proxyId:
          type: string
          example: proxy-us
        proxyConfig:
          type: string
          description: 原生代理链接（direct:// / http:// / https:// / socks5://）
        proxyBindName:
          type: string
        proxyBindUpdatedAt:
          type: string
        launchArgs:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
          example: [电商]
        keywords:
          type: array
          items:
            type: string
          example: [buyer-001]
        groupId:
          type: string
        launchCode:
          type: string
          example: BUYER_001
        running:
          type: boolean
        debugPort:
          type: integer
        debugReady:
          type: boolean
        pid:
          type: integer
        runtimeWarning:
          type: string
        lastError:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        deletedAt:
          type: string
        lastStartAt:
          type: string
        lastStopAt:
          type: string
    ProfileInput:
      type: object
      description: 创建/更新实例的配置主体
      properties:
        profileName:
          type: string
          example: buyer-001
        userDataDir:
          type: string
          description: 自定义用户数据目录；空则由系统分配
        coreId:
          type: string
          description: 绑定的内核 ID；空则使用默认内核
        fingerprintArgs:
          type: array
          items:
            type: string
          description: 指纹相关启动参数
        proxyId:
          type: string
          description: 代理池节点 ID；与 proxyConfig 同时传时优先本字段
        proxyConfig:
          type: string
          description: 自定义原生代理链接
        launchArgs:
          type: array
          items:
            type: string
          description: 实例默认启动参数
        tags:
          type: array
          items:
            type: string
        keywords:
          type: array
          items:
            type: string
        groupId:
          type: string
          description: 所属分组 ID
    LaunchRequestParams:
      type: object
      description: 仅影响本次启动的临时参数（不持久化覆盖实例配置）
      properties:
        launchArgs:
          type: array
          items:
            type: string
          description: 本次启动临时附加参数
        startUrls:
          type: array
          items:
            type: string
          description: 本次启动额外打开的网址
        skipDefaultStartUrls:
          type: boolean
          description: 是否跳过实例默认启动 URL
        proxyId:
          type: string
          description: 仅本次启动使用的代理池节点
        proxyConfig:
          type: string
          description: 仅本次启动使用的自定义代理链接
...
    ProfileWriteRequest:
      type: object
      required: [profile]
      properties:
        profile:
          $ref: '#/components/schemas/ProfileInput'
        launchCode:
          type: string
          description: 指定或覆盖实例 launchCode
          example: BUYER_001
        autoLaunch:
          type: boolean
          description: 创建/更新后是否立即启动
        start:
          $ref: '#/components/schemas/LaunchRequestParams'
...
    ProfileListResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        count:
          type: integer
          example: 1
        items:
          type: array
          items:
            $ref: '#/components/schemas/Profile'
    ProfileDetailResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        profileId:
          type: string
        profileName:
          type: string
        launchCode:
          type: string
        profile:
          $ref: '#/components/schemas/Profile'
    ProfileWriteResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        created:
          type: boolean
        updated:
          type: boolean
        launched:
          type: boolean
        profileId:
          type: string
        profileName:
          type: string
        launchCode:
          type: string
        profile:
          $ref: '#/components/schemas/Profile'
        error:
          type: string
          description: autoLaunch 失败时可能附带
    ProfileDeleteResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        deleted:
          type: boolean
          example: true
        profileId:
          type: string
        profileName:
          type: string
        launchCode:
          type: string
...
    ProfileRuntimeResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        profileId:
          type: string
        profileName:
          type: string
        launchCode:
          type: string
        running:
          type: boolean
        pid:
          type: integer
        debugPort:
          type: integer
        debugReady:
          type: boolean
        runtimeWarning:
          type: string
        lastError:
          type: string
        lastStartAt:
          type: string
        lastStopAt:
          type: string
        cdpPort:
          type: integer
        cdpUrl:
          type: string
        profile:
          $ref: '#/components/schemas/Profile'
        stopped:
          type: boolean
          description: stop 接口成功时为 true
...
paths:
...
  /api/profiles:
    get:
      tags: [Profiles]
      operationId: listProfiles
      summary: 列出全部实例
      description: 读取当前实例目录中的全部实例配置，适合做实例选择器或管理后台列表。
      responses:
        '200':
          description: 实例列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileListResponse'
              example:
                ok: true
                count: 1
                items:
                  - profileId: 550e8400-e29b-41d4-a716-446655440000
                    profileName: buyer-001
                    launchCode: BUYER_001
                    keywords: [buyer-001]
                    tags: [电商]
                    proxyId: proxy-us
                    running: false
                    debugReady: false
        '401':
          description: API Key 无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 实例目录不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [Profiles]
      operationId: createProfile
      summary: 创建实例
      description: |
        写入实例配置，可选申请 launchCode，并支持通过 `autoLaunch` 在创建后直接启动。

        - `autoLaunch=true` 时响应会附带启动结果字段
        - `proxyId` 与 `proxyConfig` 同时传时优先 `proxyId`
        - `proxyId` 无效：有 `proxyConfig` 则按自定义代理保存；否则 400
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileWriteRequest'
            examples:
              basic:
                summary: 创建并指定 launchCode
                value:
                  profile:
                    profileName: buyer-001
                    proxyId: proxy-us
                    keywords: [buyer-001]
                    tags: [电商]
                  launchCode: BUYER_001
              autoLaunch:
                summary: 创建并立即启动
                value:
                  profile:
                    profileName: buyer-002
                    keywords: [buyer-002]
                  autoLaunch: true
                  start:
                    skipDefaultStartUrls: true
              customProxy:
                summary: 自定义代理链接
                value:
                  profile:
                    profileName: buyer-003
                    proxyConfig: http://127.0.0.1:18080
                    keywords: [buyer-003]
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileWriteResponse'
        '400':
          description: 请求体非法或 profile 缺失 / 代理无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: API Key 无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: launchCode 冲突或实例数超限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 创建成功但 autoLaunch 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileWriteResponse'
  /api/profiles/{profileId}:
    parameters:
      - name: profileId
        in: path
        required: true
        schema:
          type: string
        example: 550e8400-e29b-41d4-a716-446655440000

    get:
      tags: [Profiles]
      operationId: getProfile
      summary: 查询单个实例
      description: 读取指定实例的完整配置快照。
      responses:
        '200':
          description: 实例详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileDetailResponse'
        '401':
          description: API Key 无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 实例不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [Profiles]
      operationId: updateProfile
      summary: 更新实例
      description: |
        用整份 `profile` 配置覆盖更新实例（不是 patch），可选覆盖 `launchCode`，并支持更新后立即启动。
        代理字段规则同创建接口。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileWriteRequest'
            example:
              profile:
                profileName: buyer-001-updated
                proxyId: proxy-us
                keywords: [buyer-001, checkout]
                tags: [电商]
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileWriteResponse'
        '400':
          description: 请求体非法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: API Key 无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 实例不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 更新成功但 autoLaunch 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileWriteResponse'

    delete:
      tags: [Profiles]
      operationId: deleteProfile
      summary: 删除实例
      description: 删除实例配置并移除关联 launchCode。运行中的实例会被拒绝（409），需先 stop。
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileDeleteResponse'
        '401':
          description: API Key 无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 实例不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 实例仍在运行
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/profiles/{profileId}/stop:
    post:
      tags: [Profiles]
      operationId: stopProfile
      summary: 停止指定实例
      description: 按 profileId 精确停止实例，适合任务完成后的回收。
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 停止成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileRuntimeResponse'
        '401':
          description: API Key 无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 实例不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 当前环境不支持运行态控制
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...